+++ packages/mobile/src/const/ScreenNames.enum.ts
@@ -14,4 +14,5 @@ export enum ScreenNames {
   ConnectionProcessScreen = 'ConnectionProcessScreen',
   DuplicatedUsernameScreen = 'DuplicatedUsernameScreen',
   UnregisteredUsernameScreen = 'UnregisteredUsernameScreen',
+  PossibleImpersonationAttackScreen = 'PossibleImpersonationAttackScreen',
 }